# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synops Praxis LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organisations:
    get:
      operationId: listOrganisations
      tags: [organisations]
      summary: List organisations (scoped to caller's partner)
      responses:
        "200":
          description: List of organisations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organisation"
    post:
      operationId: createOrganisation
      tags: [organisations]
      summary: Create an organisation under current partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganisationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
  /organisations/{orgId}:
    get:
      operationId: getOrganisation
      tags: [organisations]
      summary: Get organisation
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
    patch:
      operationId: updateOrganisation
      tags: [organisations]
      summary: Update organisation
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganisationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
  /organisations/{orgId}/members:
    get:
      operationId: listOrgMembers
      tags: [organisations]
      summary: List members of an organisation
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: addOrgMember
      tags: [organisations]
      summary: Add or invite a user to an organisation
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /organisations/{orgId}/stats:
    get:
      operationId: getOrgStats
      tags: [organisations]
      summary: Workforce diagnostics for org admin
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Org workforce stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgStats"

  # ── Brand Theming ──────────────────────────────────────────────────────────
...
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Modules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get module with its beats
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module with beats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /modules/{moduleId}/publish:
    post:
      operationId: publishModule
      tags: [modules]
      summary: Mark module as published (live for learners)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"

  # ── Beats ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, role]
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, partner_admin, org_admin, coach, learner]
        partnerId:
          type: ["string", "null"]
        organisationId:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Organisation:
      type: object
      required: [id, name, partnerId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        partnerId:
          type: string
        industry:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
    OrganisationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: string
    OrganisationUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
    OrgMemberInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [org_admin, coach, learner]
    OrgStats:
      type: object
      required: [orgId, totalMembers, activeEnrolments, completions, credentialsIssued, avgMasteryScore, competencyGaps]
      properties:
        orgId:
          type: string
        totalMembers:
          type: integer
        activeEnrolments:
          type: integer
        completions:
          type: integer
        credentialsIssued:
          type: integer
        avgMasteryScore:
          type: number
        competencyGaps:
          type: array
          items:
            type: string

    # ── Brand Theme ────────────────────────────────────────────────────────
...
    Module:
      type: object
      required: [id, courseId, title, status, order, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, review, published]
        order:
          type: integer
        beatCount:
          type: integer
        estimatedMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
    ModuleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        estimatedMinutes:
          type: integer
        order:
          type: integer
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, review, published]
        estimatedMinutes:
          type: integer
        order:
          type: integer
    ModuleDetail:
      type: object
      required: [id, courseId, title, status, order, beats, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        order:
          type: integer
        beats:
          type: array
          items:
            $ref: "#/components/schemas/Beat"
        estimatedMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Beats ──────────────────────────────────────────────────────────────
    Beat:
      type: object
      required: [id, moduleId, type, order, title, narration]
      properties:
        id:
          type: string
        moduleId:
          type: string
        type:
          type: string
          enum: [title_card, points, scenario, compare, diagram, close]
        order:
          type: integer
        title:
          type: string
        narration:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        scenario:
          type: ["string", "null"]
        visualData:
          description: Arbitrary JSON for template rendering
        audioUrl:
          type: ["string", "null"]
        audioStatus:
          type: string
          enum: [none, pending, ready, error]
...